requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
anyio==4.8.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.1.31
cffi==1.17.1
colorama==0.4.6
colorlog==6.9.0
discord.py==2.5.2
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
propcache==0.3.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
yarl==1.18.3
yt-dlp==2025.2.19
